Merge a main

See merge request ctd/hispanos/proyecto-integrador-2/proyecto-integrador-0723/1021ft-c1/equipo-01!75

+++ backend/api-movie/src/main/java/com/dh/movie/model/dto/movie/AllPageableDTO.java
@@ -13,5 +13,5 @@ import java.util.List;
 @NoArgsConstructor
 public class AllPageableDTO {
     private List<MovieResDTO> movies;
-    private long size;
+    private int size;
 }
